Implement functions objects in compiler and vm
Add the concept of a function object thats owns a chunk
to the compiler and vm.

Fix the chunk disassembler so it does not print out
constant data for local get/set.

Add a new object value wrapper, ObjectRef that retains the
type information of the specific ConcreteObject that
it points to.